Before Int32 to Int64 model change

+++ DocumentModel/Classes/BackstageEditBox.cs
@@ -83,7 +83,7 @@ public partial class BackstageEditBox
   /// <summary>
   /// maxLength, this property is only available in Office 2010 and later.
   /// </summary>
-  public Int32? MaxLength { get; set; }
+  public Int64? MaxLength { get; set; }
   
   /// <summary>
   /// sizeString, this property is only available in Office 2010 and later.
